requirements.txt
flask
gunicorn
flask-mail
python-dotenv
requests
reportlab
